-Rework on LPC register map (last time I promise). -Added option to manually toggle a split bank even if booting from onboard flash (TSOP recovery case when TSOP is already split). -Work on IP keypad. -Dev menu to read GPI/O state -Moved TSOPcontrol enable bit from bit1 to bit0. -IP field will be limited to 0-255 and will be rediced to 255 if something higher is entered.

+++ menu/actions/NetworkMenuActions.c
@@ -74,9 +74,9 @@ bool editIPfield(u8 * addr) {
     while(tempStringIP[cursorPos] != '\0') {
 
         if(cursorPos == 0 && tempStringIP[cursorPos] != '.')      //First byte doesn't have a '.' in before.
-            tempAddr[byteOffset++] = myAtoi(tempStringIP);
+            tempAddr[byteOffset++] = (u8)myAtoi(tempStringIP);
         else if(tempStringIP[cursorPos] == '.' && tempStringIP[cursorPos + 1] != '.' && tempStringIP[cursorPos + 1] != '\0') {       //Others do.
-            tempAddr[byteOffset++] = myAtoi(&tempStringIP[cursorPos + 1]);
+            tempAddr[byteOffset++] = (u8)myAtoi(&tempStringIP[cursorPos + 1]);
             countDots += 1;
             lastDotPos = cursorPos;
         }
@@ -92,7 +92,7 @@ bool editIPfield(u8 * addr) {
     return result;
 }
 
-u8 myAtoi(char *str)
+u16 myAtoi(char *str)
 {
     u8 i, res = 0;
